feat: put stats under "top"
closes: #890
